+++ packages/react/.storybook/main.ts
@@ -11,19 +11,20 @@ const config: StorybookConfig = {
 		getAbsolutePath("@storybook/addon-essentials"),
 		getAbsolutePath("@storybook/addon-onboarding"),
 		getAbsolutePath("@storybook/addon-interactions"),
-		{
-			name: "@storybook/addon-styling",
-			options: {
-				// Check out https://github.com/storybookjs/addon-styling/blob/main/docs/api.md
-				// For more details on this addon's options.
-				postCss: {
-					implementation: require.resolve('postcss'),
-				},
-			},
-		}
+		// getAbsolutePath("@storybook/addon-styling"),
+		// {
+		// 	name: "@storybook/addon-styling",
+		// 	options: {
+		// 		// Check out https://github.com/storybookjs/addon-styling/blob/main/docs/api.md
+		// 		// For more details on this addon's options.
+		// 		postCss: {
+		// 			implementation: require.resolve('postcss'),
+		// 		},
+		// 	},
+		// }
 	],
 	framework: {
-		name: getAbsolutePath("@storybook/react-vite"),
+		name: "@storybook/react-vite",
 		options: {},
 	},
 	docs: {
